pyproject.toml
[project]
name = "nonebot-plugin-dst-management"
version = "0.3.0"
description = "DST服务器管理NoneBot插件 - 通过DMP API管理Don't Starve Together服务器"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Xiao An", email = "admin@example.com" }
]
maintainers = [
    { name = "Xiao An", email = "admin@example.com" }
]
keywords = [
    "nonebot",
    "nonebot2",
    "nonebot-plugin",
    "don't starve together",
    "dst",
    "game-server",
    "server-management",
    "dmp-api"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Pydantic",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: Chinese (Simplified)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Typing :: Typed",
]

dependencies = [
    "nonebot2[fastapi]>=2.3.0",
    "nonebot-adapter-onebot>=2.3.0",
    "httpx>=0.24.0",
    "loguru>=0.7.0",
    "pydantic>=2.0.0",
    "lupa>=2.0.0",
]

[project.optional-dependencies]
ai = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
]
archive = [
    "zipfile36>=0.1.3",
    "python-magic>=0.4.27",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]
all = [
    "nonebot-plugin-dst-management[ai,archive,dev]",
]

[project.urls]
Homepage = "https://github.com/your-repo/nonebot-dst-management"
Documentation = "https://github.com/your-repo/nonebot-dst-management/docs"
Repository = "https://github.com/your-repo/nonebot-dst-management"
"Bug Tracker" = "https://github.com/your-repo/nonebot-dst-management/issues"
Changelog = "https://github.com/your-repo/nonebot-dst-management/blob/main/CHANGELOG.md"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[tool.pdm.scripts]
test = "pytest tests/ -v"
lint = "flake8 nonebot_plugin_dst_management/"
format = "black nonebot_plugin_dst_management/ && isort nonebot_plugin_dst_management/"
"format:check" = "black --check nonebot_plugin_dst_management/ && isort --check-only nonebot_plugin_dst_management/"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["nonebot_plugin_dst_management"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
    "@abstractmethod",
]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"
